Seperated Logic into Seperate Component File
